[level 2] Title: 최댓값과 최솟값, Time: 15.65 ms, Memory: 83.6 MB -BaekjoonHub

+++ 프로그래머스/lv2/12939. 최댓값과 최솟값/최댓값과 최솟값.java
@@ -4,8 +4,9 @@ class Solution {
     public String solution(String s) {
         String answer = "";
         
-        int[] strs = Arrays.stream(s.split(" "))
-                           .mapToInt(Integer::valueOf)
+        
+        long[] strs = Arrays.stream(s.split(" "))
+                           .mapToLong(Long::valueOf)
                            .toArray();
         
         Arrays.sort(strs);
